+++ docs/changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.10.0 - May 17, 2025
+- Replaced `dir_util.copy_tree` with `shutil.copytree` to use a built-in and maintained API in the directory merge functionality.
+- Added a test to verify the scenario of deleting paths and merging them again when using mkdocs-merge as a module.
+
 ## 0.9.0 - July 30, 2024
 - Fixed bug of `dir_util.copy_tree` caused by setuptools moving to 70.2.0 (fixes [#20](https://github.com/ovasquez/mkdocs-merge/issues/20)).
 - Updated dependency on deprecated distutils package to use setuptools version.